+++ src/toll/testing.py
@@ -1,6 +1,7 @@
 from .config import Command
 import os.path
 import pkg_resources
+import re
 import sys
 
 raw_test_command = '{0.executable} setup.py -q test'.format(sys)
@@ -18,3 +19,9 @@ def get_packages(*names):
         pkg_resources.resource_filename(
             'toll', 'fixtures/{}'.format(x)).replace(base_path, '')
         for x in names]
+
+
+def remove_ansi_codes(text):
+    """Remove ANSI colour codes from text."""
+    ansi_escape = re.compile(r'(\x9B|\x1B\[)[0-?]*[ -\/]*[@-~]')
+    return ansi_escape.sub('', text)